requirements.txt
psycopg2-binary==2.9.9
anyio==4.0.0
autopep8==2.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
pycodestyle==2.10.0
python-telegram-bot==20.6
requests==2.28.1clear
sniffio==1.3.0
urllib3==1.26.12
click==8.1.7
gTTS==2.4.0
pydub==0.25.1
python-dotenv==1.0.0
requests==2.28.1
SpeechRecognition==3.10.0
black==23.10.1
isort==5.12.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
